requirements.txt
Flask==2.0.1
cryptography==41.0.1
Flask-SocketIO==5.1.1
python-dotenv==0.19.0
PyJWT==2.3.0
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
python-socketio==5.4.0 